chore: pull and merge homepage_packages feature complete
